+++ Source/fluffy-obd-lib/diagnostics-engine/elm327/decoders/DecodeCountAB.hpp
@@ -15,10 +15,10 @@ public:
             throw std::runtime_error("invalid input");
 
         std::string byteA { text[0], text[1] };
-        float valA = Utils::hexToDec(byteA);
+        float valA = static_cast<float>(Utils::hexToDec(byteA));
         std::string byteB { text[3], text[4] };
-        float valB = Utils::hexToDec(byteB);
-        return 256 * valA + valB;
+        float valB = static_cast<float>(Utils::hexToDec(byteB));
+        return 256.0f * valA + valB;
     }
 };
